refactor: split telemetry command into subcommands and enable by default
